2.0: Move to Dune, opam files, CommonMark
Learning a lot about OCaml, this build is optimized to be reproducible
and using the modern toolchain. Dependencies are specified in an opam
file, versions are specified, it builds with dune, allows for Makefile
to do single-command testing, building, and formatting. Plays well with
opam switches.

All in all, this should allow for more hackability, so features can be
added more easily in the future.

Also introduces CommonMark, albeit not entirely self-contained. This
means we lose some ability to count words in word count, or extract
proper words in our descriptions, this looks a whole let better.