Use a cache to build incrementally.
While there may still be bugs and I don't think it has *quite* all the
features I'd like it to, we have a rudimentary cache in place to speed
up builds (previous More Pablo builds were taking 15 seconds or so).

In yet another case of "profile before you optimize," it turned out the
slowness wasn't where this change is focused (I thought it was IO in
writing out all the files) — it's from copying over the static
directory, which, with Celeste videos, was suddenly far larger than it
ever was. Given that this was a forced, *always* copy operation, we only
have to do it when it changes.

Regardless! Now *all* the code has caches in place. The code around it
is a bit of a mess, but it's working well enough for me to use on More
Pablo.

Also fixes a bug where URLs in the canonical meta tag and the sitemap
for static pages pointed to Markdown files that didn't exist.